requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
importlib-metadata==4.6.3
MouseInfo==0.1.3
Pillow==8.3.1
prettytable==2.1.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
PyTweening==1.0.3
pywhatkit==5.1
requests==2.26.0
soupsieve==2.2.1
speedtest-cli==2.1.3
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
wikipedia==1.4.0
zipp==3.5.0
